Electron setup (#1)
I used electron vite instead of forge just bc of the native integration
it has with most frameworks and it seems more maintained.

The folder is just dependency downloads, then it autobuilds an electron
repo and launches their default prebuilt scaffolding.

The way I made the .bat files is scuffed, but it was a fast and dirty
way to get it to work.

If you want me to create something more in-depth or less weird in terms
of setup lmk, I figured making everyone install node (and its
dependencies) and build a repo would be good just to make sure everyone
has the right stuff and is able to work natively in the environment
without future issue.