Feature update: Display nodes in graph within application
Simple graph display, if it looks messed up zoom in more and it'll be happy.

allowed the import of excel files through file explorer, specific file is "AirlineServerAppV7AllProd" we can make it more general and smart later, for now though this is more than enough to create and test nodes.

also deleted services and moved excelimport and neo4j to electron